Implement Unit of Work pattern

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import Message from '../_models/messages';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -39,8 +40,12 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages = this.messages.filter(message => message.id !== id);
-    });
+    this.confirmService.confirm("Confirm delete message", "This action cannot be undone").subscribe(result => {
+      if (!result) return;
+
+      this.messageService.deleteMessage(id).subscribe(() => {
+        this.messages = this.messages.filter(message => message.id !== id);
+      });
+    })
   }
 }
